requirements_AIconsole.txt
# ues python 3.8.12
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
charset-normalizer==2.1.1
cycler==0.10.0
decorator==4.4.2
dominate==2.6.0
future==0.18.2
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
jsonpatch==1.27
jsonpointer==2.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe
matplotlib==3.3.3
mkl-fft
mkl-random
mkl-service
networkx==2.5
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.46
opt-einsum==3.3.0
pandas==1.3.0
Pillow==9.2.0
protobuf==3.20
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets
pyzmq==24.0.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.17.2
scipy==1.5.4
# six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
tifffile==2020.11.26
torch==1.5.0
torchfile==0.1.0
torchvision==0.6.0
tornado==6.2
tqdm==4.54.1
urllib3==1.26.2
# visdom==0.1.8.9
websocket-client==0.57.0
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1

